refactor MarksByGroup
