+++ src/components/UI/Mark.tsx
@@ -1,6 +1,6 @@
 import { Footnote } from '@vkontakte/vkui'
 import { CSSProperties, FC, HTMLAttributes, useMemo } from 'preact/compat'
-import { getBackgroundColor, getSize, ReturnedMark } from '../../../utils'
+import { getBackgroundColor, getSize, ReturnedMark } from '../../utils'
 
 export type Sizes = 'l' | 's'
