+++ src/views/Marks.tsx
@@ -6,11 +6,11 @@ import { useEffect, useState } from 'preact/hooks'
 import PanelHeaderWithBack from '../components/UI/PanelHeaderWithBack'
 import Suspense from '../components/UI/Suspense'
 import Summary from '../components/UI/Summary'
-import MarksByGroup from '../components/MarksByGroup'
 import UserInfo from '../components/UserInfo'
 import { getPerformance } from '../methods'
 import { handleResponse, formatStatisticsData } from '../utils'
 import { useSnackbar } from '../hooks'
+import MarksByGroup from '../components/UI/Marks/MarksByGroup'
 
 const THIRD_SEC = 30 * 1000
